+++ plugin.php
@@ -5,7 +5,10 @@
  * Plugin Name:       Debug Log Manager Tool
  * Plugin URI:        https://nkb-bd.github.io/
  * Description:       Debug & Query log Helper tool with additional CLI like tools
- * Version:           3.0.5
+ * Version:           3.0.6
+ * Requires at least: 6.0
+ * Requires PHP:      7.4
+ * Tested up to:      7.0
  * Author:            Lukman Nakib
  * Author URI:        https://nkb-bd.github.io/
  * License:           GPL-2.0+
@@ -20,7 +23,7 @@ if (!defined('ABSPATH')) {
     exit;
 }
 define('DLCT_PLUGIN_URL', plugin_dir_url(__FILE__));
-define('DLCT_PLUGIN_VERSION', '3.0.5');
+define('DLCT_PLUGIN_VERSION', '3.0.6');
 define('DLCT_PLUGIN_MAIN_FILE', __FILE__);
 define('DLCT_PLUGIN_BASENAME', plugin_basename(__FILE__));
 define('DLCT_PLUGIN_DIR', plugin_dir_path(__FILE__));